Move io code into a separate module
