+++ src/io/mod.rs
@@ -1,0 +1,6 @@
+mod stdout;
+
+pub use self::stdout::{
+    stdout,
+    flush
+};
